requirements.txt
pip
Flask
cassandra-driver
jsonify
requests
